prime factors

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	int n = 612852475143;
+	long n = 612852475143;
 
 	int p;
